feat: add POST /tickets and GET /tickets
