+++ src/hooks/api/useTicket.js
@@ -2,19 +2,19 @@ import useAsync from '../useAsync';
 import * as ticketApi from '../../services/ticketApi';
 import useToken from '../useToken';
 
-export default function useTicketType() {
+export default function useTicket() {
   const token = useToken();
   const {
-    data: ticketTypes,
-    loading: ticketTypeLoading,
-    error: ticketTypeError,
-    act: getTicketTypes,
-  } = useAsync(() => ticketApi.getTicketTypes(token));
+    data: ticket,
+    loading: ticketLoading,
+    error: ticketError,
+    act: getTicket,
+  } = useAsync(() => ticketApi.getTicket(token));
 
   return {
-    ticketTypes,
-    ticketTypeLoading,
-    ticketTypeError,
-    getTicketTypes,
+    ticket,
+    ticketLoading,
+    ticketError,
+    getTicket,
   };
 }
