+++ src/services/ticketApi.js
@@ -9,4 +9,25 @@ export async function getTicketTypes(token) {
 
   return response.data;
 }
+
+export async function postTicket(ticketTypeId, token) {
+  const body = { ticketTypeId };
+  const response = await api.post('/tickets', body, {
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+  });
+
+  return response.data;
+}
+
+export async function getTicket(token) {
+  const response = await api.get('/tickets', {
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+  });
+
+  return response.data;
+}
 //
